pyproject.toml
[project]
name = "discord-bot"
version = "1.0.0"
description = "Bot Discord de la team This Is PSG"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "asyncpg==0.30.0",
    "attrs==24.3.0",
    "cffi==1.17.1",
    "discord-py==2.4.0",
    "frozenlist==1.5.0",
    "geopy==2.4.1",
    "idna==3.10",
    "multidict==6.1.0",
    "numpy==2.2.2",
    "opencv-python-headless==4.11.0.86",
    "pillow==11.1.0",
    "propcache==0.2.1",
    "pycparser==2.22",
    "pydantic>=2.0.0",
    "pynacl==1.5.0",
    "pytesseract==0.3.13",
    "python-dotenv==1.0.1",
    "yarl==1.18.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["utils", "models", "cogs"]
omit = ["tests/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = false
check_untyped_defs = true
exclude = ["tests/", "migrations/", "backup/"]
explicit_package_bases = true
